Dummy Ransomware
This is dummy randsomware that when launched will show a progress bar suggesting that it is encrypting the victims drive. 
It uses a control C class handler to ensure the victim can't easily escape the application.
This doesn't actually do anything, just thought it was funny. 
kill command will kill the process without issues.